refactor(headers): expose public types via non-detail paths
- Create types/result.h as public API for Result<T> types
- Create config/feature_flags.h as public API for feature detection
- Update all interface headers to use types/result.h
- Update public headers to use config/feature_flags.h
- Update network_system.h to use umbrella headers

This ensures public API headers don't directly include detail/ paths,
improving encapsulation while maintaining backward compatibility.

Part of #577
