+++ include/kcenon/network/concepts/concepts.h
@@ -32,7 +32,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #pragma once
 
-#include <kcenon/network/detail/config/feature_flags.h>
+#include <kcenon/network/config/feature_flags.h>
 
 /**
  * @file concepts.h
