+++ include/kcenon/network/di/service_registration.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include <kcenon/network/detail/config/feature_flags.h>
+#include <kcenon/network/config/feature_flags.h>
 
 #include <memory>
