+++ include/kcenon/network/interfaces/i_quic_client.h
@@ -61,7 +61,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <system_error>
 #include <vector>
 
-#include "kcenon/network/detail/utils/result_types.h"
+#include "kcenon/network/types/result.h"
 
 namespace kcenon::network::interfaces
 {
