+++ include/kcenon/network/network_system.h
@@ -49,11 +49,11 @@
 #include "internal/core/unified_messaging_server.h"
 
 // Protocol tags and TLS policies for unified templates
-#include "kcenon/network/detail/protocol/protocol_tags.h"
+#include "kcenon/network/protocol/protocol.h"
 #include "kcenon/network/policy/tls_policy.h"
 
 // Session management
-#include "kcenon/network/detail/session/messaging_session.h"
+#include "kcenon/network/session/session.h"
 
 // Integration interfaces
 #include "internal/integration/messaging_bridge.h"
@@ -61,12 +61,11 @@
 #include "internal/integration/container_integration.h"
 
 // Configuration system
-#include "kcenon/network/detail/config/network_config.h"
-#include "kcenon/network/detail/config/network_system_config.h"
+#include "kcenon/network/config/config.h"
 #include "internal/core/network_context.h"
 
 // Result type for error handling
-#include "kcenon/network/detail/utils/result_types.h"
+#include "kcenon/network/types/result.h"
 
 // C++20 Concepts for compile-time type validation
 #include "kcenon/network/concepts/concepts.h"
